+++ coaster/db.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from flask.ext.sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()