+++ docs/conf.py
@@ -230,9 +230,9 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  ('index', 'coaster', u'coaster Documentation',
-   u'HasGeek', 'coaster', 'Coaster contains functions and db models for recurring patterns in Flask apps.',
-   'Miscellaneous'),
+    ('index', 'coaster', u'coaster Documentation',
+     u'HasGeek', 'coaster', 'Coaster contains functions and db models for recurring patterns in Flask apps.',
+     'Miscellaneous'),
 ]
 
 # Documents to append as an appendix to all manuals.
@@ -246,4 +246,7 @@ texinfo_documents = [
 
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'http://docs.python.org/': None}
+intersphinx_mapping = {
+    'http://docs.python.org/': None,
+    'http://pythonhosted.org/Flask-SQLAlchemy/': None,
+    }